Commit 45
I created the base of GameData class

+++ Assets/Scripts/Character Scripts/EnemyScript.cs
@@ -22,7 +22,7 @@ public class EnemyScript : GeneralCharacter
     MainCharacter mainCharScript;
     [Range(1, 3)] public byte campOfEnemy;
     [HideInInspector]public byte enemyNumCode;
-    public byte enemyStaticIndex;
+    public short enemyStaticIndex;
 
     [Header("Patrol")]
     public bool hasPermanentPlace;
